+++ module/actes-generique/action/SAEEnvoiActes.class.php
@@ -87,7 +87,7 @@ class SAEEnvoiActes extends ActionExecutor {
             }
 
             $archive_path = $tmp_folder."/archive.tar.gz";
-			// ! generateArchive doit être postérieur à getBordereauNG afin que la liste des fichiers à traiter (file_list de FluxDataStandard) soit renseignée.
+			// ! generateArchive doit être postérieur à getBordereauNG afin que la liste des fichiers à traiter (file_list de FluxDataSedaDefault) soit renseignée.
 			$actesSEDA->generateArchive($fluxData,$archive_path);
 
             $donneesFormulaire->addFileFromCopy('sae_archive',"archive.tar.gz",$archive_path);