+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Rationnalisation du vocabulaire du gestionnaire de tâches (was: Démon Pastell) #708  
 - Affichage du commentaire du SAE sur l'accusé de reception et sur la réponse ainsi que de l'identiant de l'archive #815 
 - Connecteur SEDA NG : ajout de la commande size {{pastell:size:id_element_fichier}} permettant d'obtenir la taille en octet #821
+- Connecteur SEDA NG : ajout de la commande extract_zip {{pastell:extract_zip:zip_file}} permettant d'ajouter le contenu d'un fichier ZIP dans l'archive #869
 
 ## Ajouts
 
@@ -90,6 +91,7 @@ Les fonctions suivantes sont dépréciées et seront retirées dans une prochain
 - la classe Document au profit de la classe DocumentSQL
 - la table collectivite_fournisseur sera retirée dans la prochaine version
 - Pour modifier la typologie des actes, il faut passer par le champs externalData `type_piece` et plus par les champs `type_acte` et `type_pj`
+- FluxDataStandard à remplacer par FluxDataSedaDefault
 
 # [2.0.13] - 2019-06-13